separate ml part into separate module
